requirements.txt
flask==3.0.3
twilio==9.2.3
anthropic==0.30.1
python-docx==1.1.2
python-dotenv==1.0.1
